+++ Game/Scenes/GameSceneDragAndDropHandler.cs
@@ -30,7 +30,7 @@ namespace ClassicUO.Game.Scenes
                 else if (value != null && _heldItem == null)
                 {
                     UIManager.AddInputBlocker(this);
-                    UIManager.GameCursor.SetDraggedItem(value.DisplayedGraphic, value.Hue, value.Amount > 1 && value.DisplayedGraphic == value.Graphic && TileData.IsStackable((long) value.ItemData.Flags) );
+                    UIManager.GameCursor.SetDraggedItem(value.DisplayedGraphic, value.Hue, value.Amount > 1 && value.DisplayedGraphic == value.Graphic && TileData.IsStackable( value.ItemData.Flags) );
                 }
 
                 _heldItem = value;
@@ -114,7 +114,7 @@ namespace ClassicUO.Game.Scenes
             GameObject obj = SelectedObject;
             Serial serial;
 
-            if (obj is Item item && TileData.IsContainer((long) item.ItemData.Flags))
+            if (obj is Item item && TileData.IsContainer( item.ItemData.Flags))
             {
                 serial = item;
                 x = y = 0xFFFF;
